+++ src/app/stores/live.jsx
@@ -284,9 +284,10 @@ const LiveStore = Reflux.createStore({
 		this.state.live.screen.rotation = 'horizontal';
 		this.state.live.delayedRotation = 'horizontal';
 		this.state.live.battery = 100;
+		// TODO: initial state of sensors should come from live status
 		this.state.live.sensors = {};
 		this.state.live.sensors.accelerometer = {x: 0, y: 5.9, z: 0};
-		this.state.live.sensors.battery = 100;
+		this.state.live.sensors.battery = {level_percent: 100, ac_online: 1}; // eslint-disable-line camelcase
 	},
 
 	resetLive() {